+++ Smithtry1000/circuitElements.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "element.h"
+#include "LinesElement.h"
 #include "systemParameters.h"
 class CircuitElements
 {
@@ -13,10 +13,10 @@ public:
 	CircuitElements();
 	Complex z = Complex(-9999, -9999);
 	Complex y = Complex(-9999, -9999);
-	float realFirstPoint = -9999;
-	float imagFirstPoint = -9999;
-	float frequencyFirstPoint = -9999;
+	double realFirstPoint = -9999;
+	double imagFirstPoint = -9999;
+	double frequencyFirstPoint = -9999;
 	Point firstPoint;
-	map<chartMode, tuple<float, float>> chart;
+	map<chartMode, tuple<double, double>> chart;
 	~CircuitElements();
 };
